Ignore pdftex errors in the Makefile, because
LaTeX is getting upset that pandoc is putting ∃ in its sources.  I think
that's a bug in either pandoc, or the LaTeX unicode packages, actually, since
it seems pandoc is including this in its LaTeX output:

\ifnum 0\ifxetex 1\fi\ifluatex 1\fi=0 % if pdftex
  \usepackage[utf8]{inputenc}
\else % if luatex or xelatex
  \usepackage{fontspec}
  \ifxetex
    \usepackage{xltxtra,xunicode}
  \fi
  \defaultfontfeatures{Mapping=tex-text,Scale=MatchLowercase}
  \newcommand{\euro}{€}
\fi

That's not enough to get Unicode math symbols working I guess.  I should
investigate later whether this is a bug in the pandoc template (i.e., ∃ a
LaTeX package that supports math symbols via Unicode, but pandoc isn't using
it), a failure for LaTeX to provide any such package, in which case it's
really pandoc's failure to convert ∃ to $\exists$ or something like that.