Update to recent changes of ASF website.
Since I last gave this talk, ASF changed the website.  Although not very
much -- they changed it just enough so that it's more difficult to
criticize, which a mere political move.

ASF should take it down if they really aren't anti-copyleft.  It still
insinuates a connection between the BSA and copyleft enforcement, and it
now says that copyleft makes people ill.

I'll make this point in my talk.